+++ tests/live_tests/scenarios/communities/base.py
@@ -268,3 +268,9 @@ class BGPCommunitiesScenarioOpenBGPD62(BGPCommunitiesScenarioOpenBGPD):
     __test__ = False
 
     TARGET_VERSION = "6.2"
+
+
+class BGPCommunitiesScenarioOpenBGPD63(BGPCommunitiesScenarioOpenBGPD):
+    __test__ = False
+
+    TARGET_VERSION = "6.3"