+++ tests/live_tests/scenarios/default/base.py
@@ -83,7 +83,8 @@ class DefaultConfigScenarioOpenBGPD(LiveScenario_TagRejectPolicy,
             cls.DATA["rs_IPAddress"],
             [
                 (
-                    cls.build_rs_cfg("openbgpd", "main.j2", "rs.conf", None),
+                    cls.build_rs_cfg("openbgpd", "main.j2", "rs.conf", None,
+                                     target_version="6.0"),
                     "/etc/bgpd.conf"
                 )
             ]