+++ tests/live_tests/scenarios/tag_as_set/base.py
@@ -728,3 +728,8 @@ class TagASSetScenarioOpenBGPD62(TagASSetScenarioOpenBGPD):
     __test__ = False
 
     TARGET_VERSION = "6.2"
+
+class TagASSetScenarioOpenBGPD63(TagASSetScenarioOpenBGPD):
+    __test__ = False
+
+    TARGET_VERSION = "6.3"